Time: 14 ms (8.45%) | Memory: 18.7 MB (6.13%) - LeetSync

+++ 662-maximum-width-of-binary-tree/maximum-width-of-binary-tree.cpp
@@ -18,27 +18,27 @@ public:
 
         queue<pair<TreeNode*,long long int>>q;
         q.push({root,0});
-        long long int width =0;
+         int width =0;
         while(!q.empty()){
 
             int size = q.size();
-            long long int min_index = q.front().second;
-            long long int first_index,last_index;
+             int min_index = q.front().second;
+            int first_index,last_index;
 
             for(int i=0;i<size;i++){
                 TreeNode* temp = q.front().first;
 
-                long long int curr = q.front().second - min_index;
+                int curr = q.front().second - min_index;
                 q.pop();
 
                 if(i==0) first_index = curr;
                 if(i==size-1) last_index = curr;
 
                 if(temp->left){
-                    q.push({temp->left , 2*curr +1});
+                    q.push({temp->left ,(long long )2*curr +1});
                 }
                 if(temp->right){
-                    q.push({temp->right , 2*curr +2});
+                    q.push({temp->right , (long long )2*curr +2});
                 }
             }
             width = max(width , last_index - first_index + 1);
